+++ src/main/java/com/soklet/internal/microhttp/Logger.java
@@ -13,6 +13,13 @@ public interface Logger {
     boolean enabled();
 
     /**
+     * Indicates whether failure logging is active.
+     */
+    default boolean failureEnabled() {
+        return enabled();
+    }
+
+    /**
      * Creates a new log event consisting of multiple log entries.
      */
     void log(LogEntry... entries);
@@ -21,6 +28,20 @@ public interface Logger {
      * Creates a new log event consisting of an exception and multiple log entries.
      */
     void log(Exception e, LogEntry... entries);
+
+    /**
+     * Creates a new failure log event consisting of multiple log entries.
+     */
+    default void logFailure(LogEntry... entries) {
+        log(entries);
+    }
+
+    /**
+     * Creates a new failure log event consisting of an exception and multiple log entries.
+     */
+    default void logFailure(Exception e, LogEntry... entries) {
+        log(e, entries);
+    }
 
     /**
      * Creates a new log event consisting of a throwable and multiple log entries.
@@ -32,6 +53,18 @@ public interface Logger {
         }
 
         log(new RuntimeException(throwable), entries);
+    }
+
+    /**
+     * Creates a new failure log event consisting of a throwable and multiple log entries.
+     */
+    default void logFailure(Throwable throwable, LogEntry... entries) {
+        if (throwable instanceof Exception e) {
+            logFailure(e, entries);
+            return;
+        }
+
+        logFailure(new RuntimeException(throwable), entries);
     }
 
 }